// src/features/movies/types/MovieInterface.ts
export interface BelongsToCollection {
  id?: number;
  name?: string;
  poster_path?: string;
  backdrop_path?: null;
}

export interface Genre {
  id: number;
  name: string;
}

export interface ProductionCompany {
  id: number;
  logo_path: null | string;
  name: string;
  origin_country: string;
}

export interface ProductionCountry {
  iso_3166_1: string;
  name: string;
}

export interface SpokenLanguage {
  english_name: string;
  iso_639_1: string;
  name: string;
}

export interface ListMovieDetailsDataResponse {
  adult?: boolean;
  backdrop_path?: string;
  belongs_to_collection?: BelongsToCollection;
  budget?: number;
  genres?: Genre[];
  homepage?: string;
  id?: number;
  imdb_id?: string;
  original_language?: string;
  original_title?: string;
  overview?: string;
  popularity?: number;
  poster_path?: string;
  production_companies?: ProductionCompany[];
  production_countries?: ProductionCountry[];
  release_date?: Date;
  revenue?: number;
  runtime?: number;
  spoken_languages?: SpokenLanguage[];
  status?: string;
  tagline?: string;
  title?: string;
  video?: boolean;
  vote_average?: number;
  vote_count?: number;
}

export interface ResultMovieDetailsData {
  data?: ListMovieDetailsDataResponse;
}

export interface ListMoviesResponse<T> {
  page?: number;
  results?: T[];
  total_pages?: number;
  total_results?: number;
}

export interface Results {
  adult?: boolean;
  backdrop_path?: null | string;
  genre_ids?: number[];
  id: number;
  original_language?: string;
  original_title?: string;
  overview?: string;
  popularity?: number;
  poster_path: null | string;
  release_date?: string;
  title: string;
  video?: boolean;
  vote_average?: number;
  vote_count?: number;
  onListItemClick?: (movieId: number) => void;
}

export interface ResultListDiscoverData {
  data?: ListMoviesResponse<Results>;
  moviesDiscoverCurrentList?: Results[];
  onListItemClick?: (movieId: number) => void;
  isLoadingMoviesDiscover?: boolean;
}